refactor: split test & perf logic into separate files
